kinoko_takenoko/requirements.txt
pillow
chainer
argparse
sklearn
matplotlib
